tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/005.html

<!--
var x;
var y;
try {
  importScripts('data:text/javascript,x={',
                'data:text/javascript,}');
} catch(e) {
  y = true;
}
postMessage([x, y]);
/*
-->
<!doctype html>
<title>importScripts separate scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data[0], undefined);
    assert_true(e.data[1]);
    this.done();
  });
});
</script>
<!--
*/
//-->
